make _demo runnable by itself
- go run main.go
