+++ _demo/demo/fonts.go
@@ -1,4 +1,4 @@
-package main
+package demo
 
 import (
 	"io/ioutil"
