+++ _demo/demo/page.go
@@ -1,4 +1,4 @@
-package main
+package demo
 
 import (
 	"fmt"
