+++ _demo/demo/resources.go
@@ -1,4 +1,4 @@
-package main
+package demo
 
 import (
 	"image/color"
