+++ _demo/demo/sizedpanel.go
@@ -1,4 +1,4 @@
-package main
+package demo
 
 import (
 	"image"
