+++ _demo/demo/tooltip.go
@@ -1,4 +1,4 @@
-package main
+package demo
 
 import (
 	"time"
