+++ _demo/go.mod
@@ -11,3 +11,5 @@ require (
 	golang.org/x/image v0.0.0-20200927104501-e162460cd6b5
 	golang.org/x/text v0.3.4
 )
+
+replace github.com/blizzy78/ebitenui/_demo/demo v0.0.0 => ./_demo/demo/
